@@ -1,17 +1,18 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
   </PropertyGroup>
-  
+
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>  
-  
+  </ItemGroup>
+
   <ItemGroup>
-    <None Remove="bin\Debug\netcoreapp3.1\Codebot.Web.deps.json" />
-    <None Remove="bin\Debug\netcoreapp3.1\Codebot.Web.dll" />
-    <None Remove="bin\Debug\netcoreapp3.1\Codebot.Web.pdb" />
+    <None Remove="bin\Debug\net6.0\Codebot.Web.deps.json" />
+    <None Remove="bin\Debug\net6.0\Codebot.Web.dll" />
+    <None Remove="bin\Debug\net6.0\Codebot.Web.pdb" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Codebot.Web.csproj" />
@@ -21,20 +22,20 @@
     <None Include="obj\project.nuget.cache" />
     <None Include="obj\project.assets.json" />
     <None Include="obj\Codebot.Web.csproj.nuget.g.props" />
-    <None Include="obj\Debug\netcoreapp3.1\Codebot.Web.csprojAssemblyReference.cache" />
-    <None Include="obj\Debug\netcoreapp3.1\Codebot.Web.csproj.CoreCompileInputs.cache" />
-    <None Include="obj\Debug\netcoreapp3.1\Codebot.Web.dll" />
-    <None Include="obj\Debug\netcoreapp3.1\Codebot.Web.assets.cache" />
-    <None Include="obj\Debug\netcoreapp3.1\Codebot.Web.AssemblyInfoInputs.cache" />
-    <None Include="obj\Debug\netcoreapp3.1\Codebot.Web.csproj.FileListAbsolute.txt" />
-    <None Include="obj\Debug\netcoreapp3.1\Codebot.Web.pdb" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="obj\Debug\netcoreapp3.1\Codebot.Web.AssemblyInfo.cs" />
+    <None Include="obj\Debug\net6.0\Codebot.Web.csprojAssemblyReference.cache" />
+    <None Include="obj\Debug\net6.0\Codebot.Web.csproj.CoreCompileInputs.cache" />
+    <None Include="obj\Debug\net6.0\Codebot.Web.dll" />
+    <None Include="obj\Debug\net6.0\Codebot.Web.assets.cache" />
+    <None Include="obj\Debug\net6.0\Codebot.Web.AssemblyInfoInputs.cache" />
+    <None Include="obj\Debug\net6.0\Codebot.Web.csproj.FileListAbsolute.txt" />
+    <None Include="obj\Debug\net6.0\Codebot.Web.pdb" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="src\security\" />
     <Folder Include="src\xml\" />
     <Folder Include="src\common\" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="system.net.http" Version="4.3.4" />
   </ItemGroup>
 </Project>
