@@ -1,18 +1,17 @@
-﻿using System;
+﻿namespace Codebot.Web;
+
+using System;
 using System.Reflection;
 
-namespace Codebot.Web
+public static class AttributeExtensions
 {
-    public static class AttributeExtensions
+    public static T GetCustomAttribute<T>(this MemberInfo memberInfo, bool inherit = false) where T : Attribute
     {
-        public static T GetCustomAttribute<T>(this MemberInfo memberInfo, bool inherit = false) where T : Attribute
-        {
-            return (T)Attribute.GetCustomAttribute(memberInfo, typeof(T), inherit);
-        }
+        return (T)Attribute.GetCustomAttribute(memberInfo, typeof(T), inherit);
+    }
 
-        public static T[] GetCustomAttributes<T>(this MemberInfo memberInfo, bool inherit = false) where T : Attribute
-        {
-            return (T[])Attribute.GetCustomAttributes(memberInfo, typeof(T), inherit);
-        }
+    public static T[] GetCustomAttributes<T>(this MemberInfo memberInfo, bool inherit = false) where T : Attribute
+    {
+        return (T[])Attribute.GetCustomAttributes(memberInfo, typeof(T), inherit);
     }
 }
