@@ -1,21 +1,20 @@
-﻿using System;
+﻿namespace Codebot.Web;
+
+using System;
 using Microsoft.AspNetCore.Http;
 
-namespace Codebot.Web
+public class ContextEventArgs
 {
-    public class ContextEventArgs
-    {
-        public HttpContext Context { get; set; }
-        public bool Handled { get; set; }
-        public IHttpHandler Handler { get; set; }
-        public Exception Error { get; set; }
+    public HttpContext Context { get; set; }
+    public bool Handled { get; set; }
+    public IHttpHandler Handler { get; set; }
+    public Exception Error { get; set; }
 
-        public ContextEventArgs(HttpContext context, Exception error = null)
-        {
-            Context = context;
-            Error = error;
-            Handler = null;
-            Handled = false;
-        }
+    public ContextEventArgs(HttpContext context, Exception error = null)
+    {
+        Context = context;
+        Error = error;
+        Handler = null;
+        Handled = false;
     }
 }
