@@ -1,19 +1,18 @@
-﻿using System;
+﻿namespace Codebot.Web;
 
-namespace Codebot.Web
+using System;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class PageTypeAttribute : Attribute
 {
-    [AttributeUsage(AttributeTargets.Class)]
-    public class PageTypeAttribute : Attribute
+    public PageTypeAttribute(string fileName)
     {
-        public PageTypeAttribute(string fileName)
-        {
-            ContentType = "text/html; charset=utf-8";
-            FileName = fileName;
-            IsTemplate = false;
-        }
-
-        public string ContentType { get; set; }
-        public string FileName { get; set; }
-        public bool IsTemplate { get; set; }
+        ContentType = "text/html; charset=utf-8";
+        FileName = fileName;
+        IsTemplate = false;
     }
+
+    public string ContentType { get; set; }
+    public string FileName { get; set; }
+    public bool IsTemplate { get; set; }
 }
